+++ cmd/termshark/termshark.go
@@ -64,12 +64,51 @@ func main() {
 
 	res := cmain()
 	ensureGoroutinesStopWG.Wait()
+
+	if !termshark.ShouldSwitchTerminal && !termshark.ShouldSwitchBack {
+		os.Exit(res)
+	}
+
+	os.Clearenv()
+	for _, e := range termshark.OriginalEnv {
+		ks := strings.SplitN(e, "=", 2)
+		if len(ks) == 2 {
+			os.Setenv(ks[0], ks[1])
+		}
+	}
+
+	exe, err := os.Executable()
+	if err != nil {
+		log.Warnf("Unexpected error determining termshark executable: %v", err)
+		os.Exit(1)
+	}
+
+	switch {
+	case termshark.ShouldSwitchTerminal:
+		os.Setenv("TERMSHARK_ORIGINAL_TERM", os.Getenv("TERM"))
+		os.Setenv("TERM", fmt.Sprintf("%s-256color", os.Getenv("TERM")))
+	case termshark.ShouldSwitchBack:
+		os.Setenv("TERM", os.Getenv("TERMSHARK_ORIGINAL_TERM"))
+		os.Setenv("TERMSHARK_ORIGINAL_TERM", "")
+	}
+
+	// Need exec because we really need to re-initialize everything, including have
+	// all init() functions be called again
+	err = syscall.Exec(exe, os.Args, os.Environ())
+	if err != nil {
+		log.Warnf("Unexpected error exec-ing termshark %s: %v", exe, err)
+		res = 1
+	}
+
 	os.Exit(res)
 }
 
 func cmain() int {
 	startedSuccessfully := false // true if we reached the point where packets were received and the UI started.
 	uiSuspended := false         // true if the UI was suspended due to SIGTSTP
+
+	// Preserve in case we need to re-exec e.g. if the user switches TERM
+	termshark.OriginalEnv = os.Environ()
 
 	sigChan := make(chan os.Signal, 100)
 	// SIGINT and SIGQUIT will arrive only via an external kill command,
@@ -275,7 +314,7 @@ func cmain() int {
 	// terminal emumlator supports 256 colors.
 	termVar := termshark.ConfString("main.term", "")
 	if termVar != "" {
-		log.Infof("Configuration file overrides TERM setting, using TERM=%s", termVar)
+		fmt.Fprintf(os.Stderr, "Configuration file overrides TERM setting, using TERM=%s\n", termVar)
 		os.Setenv("TERM", termVar)
 	}
 
@@ -1098,8 +1137,9 @@ Loop:
 			// Need to do that here because the app won't know how many colors the screen
 			// has (and therefore which variant of the theme to load) until the screen is
 			// activated.
-			mode := theme.Mode(app.GetColorMode()).String() // more concise
-			themeName := termshark.ConfString(fmt.Sprintf("main.theme-%s", mode), "default")
+			mode := app.GetColorMode()
+			modeStr := theme.Mode(mode) // more concise
+			themeName := termshark.ConfString(fmt.Sprintf("main.theme-%s", modeStr), "default")
 			loaded := false
 			if themeName != "" {
 				err = theme.Load(themeName, app)
@@ -1131,6 +1171,33 @@ Loop:
 			startedSuccessfully = true
 
 			ui.StartUIChan = nil // make sure it's not triggered again
+
+			if runtime.GOOS != "windows" {
+				if mode == gowid.Mode8Colors {
+					// If exists is true, it means we already tried and then reverted back, so
+					// just load up termshark normally with no further interruption.
+					if _, exists := os.LookupEnv("TERMSHARK_ORIGINAL_TERM"); !exists {
+						if !termshark.ConfBool("main.disable-term-helper", false) {
+							err = termshark.Does256ColorTermExist()
+							if err != nil {
+								log.Infof("Must use 8-color mode because 256-color version of TERM=%s unavailable - %v.", os.Getenv("TERM"), err)
+							} else {
+								time.AfterFunc(time.Duration(3)*time.Second, func() {
+									app.Run(gowid.RunFunction(func(app gowid.IApp) {
+										ui.SuggestSwitchingTerm(app)
+									}))
+								})
+							}
+						}
+					}
+				} else if os.Getenv("TERMSHARK_ORIGINAL_TERM") != "" {
+					time.AfterFunc(time.Duration(3)*time.Second, func() {
+						app.Run(gowid.RunFunction(func(app gowid.IApp) {
+							ui.IsTerminalLegible(app)
+						}))
+					})
+				}
+			}
 
 			defer func() {
 				// Do this to make sure the program quits quickly if quit is invoked