+++ utils.go
@@ -40,6 +40,7 @@ import (
 	"github.com/gcla/termshark/v2/system"
 	"github.com/gcla/termshark/v2/widgets/resizable"
 	"github.com/gdamore/tcell"
+	"github.com/gdamore/tcell/terminfo/dynamic"
 	"github.com/mattn/go-isatty"
 	"github.com/pkg/errors"
 	"github.com/shibukawa/configdir"
@@ -99,10 +100,13 @@ var InternalErr = InternalError{}
 //======================================================================
 
 var (
-	UserGuideURL string = "https://termshark.io/userguide"
-	FAQURL       string = "https://termshark.io/faq"
-	BugURL       string = "https://github.com/gcla/termshark/issues/new?assignees=&labels=&template=bug_report.md&title="
-	FeatureURL   string = "https://github.com/gcla/termshark/issues/new?assignees=&labels=&template=feature_request.md&title="
+	UserGuideURL         string = "https://termshark.io/userguide"
+	FAQURL               string = "https://termshark.io/faq"
+	BugURL               string = "https://github.com/gcla/termshark/issues/new?assignees=&labels=&template=bug_report.md&title="
+	FeatureURL           string = "https://github.com/gcla/termshark/issues/new?assignees=&labels=&template=feature_request.md&title="
+	OriginalEnv          []string
+	ShouldSwitchTerminal bool
+	ShouldSwitchBack     bool
 )
 
 //======================================================================
@@ -1320,6 +1324,11 @@ func FileSizeDifferentTo(filename string, cur int64) (int64, bool) {
 		}
 	}
 	return newSize, diff
+}
+
+func Does256ColorTermExist() error {
+	_, _, e := dynamic.LoadTerminfo(fmt.Sprintf("%s-256color", os.Getenv("TERM")))
+	return e
 }
 
 //======================================================================